+++ projects/ng-gallery/src/lib/components/gallery-nav.component.ts
@@ -1,5 +1,6 @@
-import { Component, OnInit, Input, Output, EventEmitter, ChangeDetectionStrategy } from '@angular/core';
+import { Component, OnInit, Input, ChangeDetectionStrategy } from '@angular/core';
 import { DomSanitizer, SafeHtml } from '@angular/platform-browser';
+import { Gallery } from '../services/gallery.service';
 import { GalleryState } from '../models/gallery.model';
 import { GalleryConfig } from '../models/config.model';
 
@@ -11,25 +12,25 @@ import { GalleryConfig } from '../models/config.model';
        class="g-nav-prev"
        aria-label="Previous"
        role="button"
-       (click)="action.emit('prev')"
+       (click)="gallery.ref(this.galleryId).prev()"
        [innerHtml]="navIcon"></i>
 
     <i *ngIf="config.loop || state.hasNext"
        class="g-nav-next"
        aria-label="Next"
        role="button"
-       (click)="action.emit('next')"
+       (click)="gallery.ref(this.galleryId).next()"
        [innerHtml]="navIcon"></i>
   `
 })
 export class GalleryNavComponent implements OnInit {
 
   navIcon: SafeHtml;
+  @Input() galleryId: string;
   @Input() state: GalleryState;
   @Input() config: GalleryConfig;
-  @Output() action = new EventEmitter<string>();
 
-  constructor(private _sanitizer: DomSanitizer) {
+  constructor(public gallery: Gallery, private _sanitizer: DomSanitizer) {
   }
 
   ngOnInit() {