+++ projects/ng-gallery/src/lib/models/config.model.ts
@@ -25,6 +25,7 @@ export interface GalleryConfig {
   thumbSlidingDisabled?: boolean;
   mouseSlidingDisabled?: boolean;
   thumbMouseSlidingDisabled?: boolean;
+  contentVisibilityAuto?: boolean;
   resizeDebounceTime?: number;
   playerInterval?: number;
   boxTemplate?: TemplateRef<any>;