+++ projects/ng-gallery/src/lib/utils/gallery.default.ts
@@ -33,6 +33,7 @@ export const defaultConfig: GalleryConfig = {
   thumbWidth: 120,
   thumbHeight: 90,
   disableThumb: false,
+  contentVisibilityAuto: true,
   slidingDisabled: false,
   thumbSlidingDisabled: false,
   mouseSlidingDisabled: false,